// src/testData/interfaces/VehicleDetails.ts
export interface VehicleDetails {
    ModelName: string;
    ModelYear?: string;
    Price: string;
    KM?: string;
    FuelType?: string;
    ANCAPRating?: number;
    BodyType?: string;
    Seats?: number;
    Transmission?: string;
    VIN?: string;
    Cylinders?: number;
    Reg?: string;
    Colour?: string;
    Upholstery?: string;
    Doors?: number;
    EngineCapacity?: number;
    Typeclass?: string;
    Gears?: number;
    RegExpiry?: string;

}